User words file is unexpectedly removed from disk
PDFOC-93

+++ itext/itext.pdfocr.tesseract4/itext/pdfocr/tesseract4/Tesseract4ExecutableOcrEngine.cs
@@ -184,7 +184,8 @@ namespace iText.Pdfocr.Tesseract4 {
                         , imagePath, e.Message));
                 }
                 try {
-                    if (GetTesseract4OcrEngineProperties().GetPathToUserWordsFile() != null) {
+                    if (GetTesseract4OcrEngineProperties().GetPathToUserWordsFile() != null && GetTesseract4OcrEngineProperties
+                        ().IsUserWordsFileTemporary()) {
                         TesseractHelper.DeleteFile(GetTesseract4OcrEngineProperties().GetPathToUserWordsFile());
                     }
                 }
