+++ itext/itext.pdfocr.tesseract4/itext/pdfocr/tesseract4/Tesseract4LibOcrEngine.cs
@@ -179,7 +179,8 @@ namespace iText.Pdfocr.Tesseract4 {
                 if (tesseractInstance != null) {
                     TesseractOcrUtil.DisposeTesseractInstance(tesseractInstance);
                 }
-                if (GetTesseract4OcrEngineProperties().GetPathToUserWordsFile() != null) {
+                if (GetTesseract4OcrEngineProperties().GetPathToUserWordsFile() != null && GetTesseract4OcrEngineProperties
+                    ().IsUserWordsFileTemporary()) {
                     TesseractHelper.DeleteFile(GetTesseract4OcrEngineProperties().GetPathToUserWordsFile());
                 }
             }
